django_image_compressor/tests/requirements.txt
asgiref==3.3.1
Django==3.1.3
django-image-compressor==0.3.5
Pillow==8.0.1
pytz==2020.4
sqlparse==0.4.1
